Tier-3: dedup chunking + CLAUDE.md file/type layout (B4, C1-C4)
B4: CategoryLoadResult.casksCoupled reimplemented Array.chunked(into:) — call the
existing helper.

Conventions (one view/type per file; owned types out of Extensions/):
- C1: extract BrokenInstallView from DetailView.swift into its own file.
- C2: extract EnvironmentInput from MirrorsView.swift into Components/.
- C3: move View+CardActionPill.swift (owned modifier + button style) from
  Extensions/ to Components/CardActionPill.swift.
- C4: rename FontExtension.swift → Font+Applite.swift.

Build succeeds.
